Moved output files from biohansel and phylogenetic tree view and replaced with output files tab
