+++ src/main/webapp/resources/js/pages/analysis/components/Analysis.jsx
@@ -95,7 +95,7 @@ export default function Analysis() {
         } else if (analysisContext.analysisType === "BIO_HANSEL") {
           tabLinks.push(
             <Menu.Item key="biohansel">
-              <Link to={`${DEFAULT_URL}/${ANALYSIS.BIOHANSEL}/`}>
+              <Link to={`${DEFAULT_URL}/${ANALYSIS.BIOHANSEL}`}>
                 {i18n("Analysis.biohansel")}
               </Link>
             </Menu.Item>
@@ -106,7 +106,7 @@ export default function Analysis() {
         ) {
           tabLinks.push(
             <Menu.Item key="tree">
-              <Link to={`${DEFAULT_URL}/${ANALYSIS.TREE}/`}>
+              <Link to={`${DEFAULT_URL}/${ANALYSIS.TREE}`}>
                 {i18n("Analysis.phylogeneticTree")}
               </Link>
             </Menu.Item>
